Extract method allocation analysis
Move allocation occurrence discovery, escape refinement, and allocation-specific control-flow interpretation behind a focused per-method service while preserving shared decode and builder orchestration.
